Moving NullLoggerFactory into an internal namespace
